Adding exec role to support installing/updating packages
The below role can be used to upgrade/install new packages or binaries
using predefined shell scripts built in the hostconfig-operator.
The script name has to be specified in the CR so that it can be used to
perform required configuration on the nodes. The scripts can be
customized to perform security upgrades or any other configuration
required on the host.

Change-Id: I4400d2f278749082805bb1131bf2c3bc7f95fadb